Added sanitized input parameters, CSRTrack, Variable Bunch Compressor angles
Former-commit-id: a0a5eb5e84de772f1840a2c50b0327235860402a

+++ ASTRAFramework/Simple/CSRTrack.py
@@ -24,9 +24,10 @@ class CSRTrack(object):
         self.CSRTrackCommand = command
 
     def runCSRTrackFile(self, filename):
-        command = self.CSRTrackCommand + [filename]
-        with open(os.devnull, "w") as f:
-            subprocess.call(command, stdout=f, cwd=self.subdir)
+        if self.overwrite:
+            command = self.CSRTrackCommand + [filename]
+            with open(os.devnull, "w") as f:
+                subprocess.call(command, stdout=f, cwd=self.subdir)
 
     def csrtrackinputtext(self, ANGLE):
         return """io_path{logfile = log.txt}
